requirements.txt
torch
pillow
tqdm
torchvision
lmdb